Fixed UniversalBootstrap to be able to check if we are using Network.Framework without specifying
